--- packages/pages-asset-server/.gitignore
@@ -1,1 +1,0 @@
-/lib
