--- packages/pages-asset-server/README.md
@@ -1,10 +1,0 @@
-# @cloudflare/pages-asset-server
-
-This small utility package is used by `wrangler pages dev` to serve static files of a directory.
-
-## To Do List
-
-- [x] esbuild-jest
-- [ ] esbuild instead of tsc
-- [ ] Auto-releases with GitHub Actions
-- [ ] Figure out what we want to do with cache headers
