--- packages/pages-asset-server/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "@cloudflare/pages-asset-server",
-  "version": "0.0.0",
-  "bin": "./lib/index.js",
-  "type": "module",
-  "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
-  "license": "MIT OR Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/cloudflare/wrangler/issues"
-  },
-  "homepage": "https://github.com/cloudflare/wrangler2/tree/main/packages/pages-asset-server#readme",
-  "devDependencies": {
-    "@types/express": "^4.17.13",
-    "@types/yargs": "^17.0.5",
-    "execa": "^5.1.1",
-    "express": "^4.17.1",
-    "rimraf": "^3.0.2",
-    "serve-static": "^1.14.1",
-    "typescript": "^4.4.4",
-    "yargs": "^17.2.1"
-  },
-  "files": [
-    "/lib",
-    "/README.md"
-  ],
-  "scripts": {
-    "clean": "rimraf lib",
-    "prepack": "npm run clean && npm run bundle",
-    "bundle": "tsc && chmod +x ./lib/index.js",
-    "build": "npm run prepack",
-    "start": "npm run bundle && NODE_OPTIONS=--enable-source-maps ./lib/index.js",
-    "test": "npm run build && NODE_OPTIONS=--enable-source-maps jest --rootDir src --silent=false"
-  },
-  "engines": {
-    "node": ">=16.7.0"
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.tsx?$": [
-        "esbuild-jest",
-        {
-          "sourcemap": true
-        }
-      ]
-    }
-  },
-  "dependencies": {
-    "chokidar": "^3.5.2",
-    "cross-fetch": "^3.1.4"
-  }
-}
