--- packages/pages-asset-server/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "importHelpers": true,
-    "esModuleInterop": true,
-    "moduleResolution": "node",
-    "outDir": "lib",
-    "target": "ES2020",
-    "strict": true
-  },
-  "include": ["src/**/*"],
-  "exclude": ["**/*.test.ts"]
-}
